feat: Add methods for handling request headers and ResponseEntity
Added two new POST methods to the FirstController: "headersParamsMethod" and "headersListParamsMethod" to handle request headers. Utilized @RequestHeader annotation to receive header values, and @RequestHeader Map<String, String> to receive all headers as a map.

Also, implemented a new GET method "responseEntityMethod" to demonstrate the usage of ResponseEntity. The method returns different responses based on the value of the path variable "id", with HttpStatus.OK for id > 5 and a bad request otherwise.