feat: Implement Inversion of Control (IoC) and Dependency Injection (DI)
Added a new REST controller `MyControllerComponent` in the package `br.com.gufvr.ioc_di` to demonstrate Inversion of Control (IoC) and Dependency Injection (DI). Autowired the `MyComponent` bean to the controller, and implemented a GET mapping to call the component's method. Created a service class `MyComponent` annotated with @Service to be managed by Spring's IoC container.